add shortname and production hostname add task to convert old schema to new
